// src/app/features/guest/models/guest-accreditation.model.ts
/*
 * Copyright 2021 Bundesrepublik Deutschland
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import GuestExtendedFormModel from './guest-extended-form.model';

/**
 * Model containing the guest accreditation attributes
 */
export default class GuestAccreditionModel {
  connectionQrCode: string;

  guest: GuestExtendedFormModel;

  id: string;

  invitationEmail: string;

  invitationLink: string;

  status: string;
}